requirements.txt
anthropic>=0.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
chromadb>=0.5.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-anthropic>=0.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-dotenv>=1.0.0
gitpython>=3.1.0
tabulate>=0.9.0
rich>=13.0.0
